# api.yaml
openapi: 3.0.0
info:
  title: ZenCura Hospital Management System API
  description: API documentation for the ZenCura HMS V2, allowing for patient management, appointment scheduling, and clinical reporting.
  version: 2.0.0
servers:
  - url: http://localhost:5000
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /api/doctor/appointments:
    get:
      summary: Fetch assigned appointments for a Doctor
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of scheduled visits
...